Create ProbabilityAlgo
This algorithm assigns word - score according to no of most probable words occurring in it. 
Start with AROUSE - highest probability . It then asks for letters that were found correct and letters that are not in the word and suggests the hints.
Try it . It wont give you the complete direct answers but only small hints.